Persistence tests are now only being run on nightly
