+++ .travis.yml
@@ -8,7 +8,7 @@ rust:
 
 matrix:
   allow_failures:
-    - rust: stable
+    - rust: nightly
   fast_finish: true
 
 script:
