+++ src/persistence/src/lib.rs
@@ -16,36 +16,14 @@
   along with the Purple Library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#![feature(custom_test_frameworks, test)]
-#![test_runner(test_runner)]
-
 #[macro_use] extern crate bin_tools;
 
 extern crate byteorder;
-extern crate test;
 extern crate ansi_term;
 
-use test::{TestName, TestFn, TestDescAndFn};
 use ansi_term::Colour::Green;
 use std::path::PathBuf;
 use rocksdb::DB;
-
-fn test_runner(tests: &[&TestDescAndFn]) {
-    for t in tests {
-        if let TestFn::StaticTestFn(fun) = t.testfn {
-            fun();
-            let test_name = if let TestName::StaticTestName(name) = t.desc.name {
-                name
-            } else {
-                panic!("No static test name");
-            };
-
-            println!("test {} ... {}", test_name, Green.paint("ok"));
-        } else {
-            panic!("");
-        }
-    }
-}
 
 pub fn open_database(path: &PathBuf) -> DB {
     DB::open(&crate::db_options(), path.to_str().unwrap()).unwrap()
